requirements.txt
numpy==1.24.4
opencv_contrib_python==4.10.0.84
Pillow==11.0.0
pytorch_msssim==1.0.0
scipy==1.11.4
skimage
torch==2.3.1+cu118
torchsummary==1.5.1
torchvision==0.18.1+cu118
torchvision==0.18.1